Adição de Cliente.hpp ao diretório include
